Added got status and got reset (WIP)
At the moment got status simply prints the current version, later
it can report more info as we see fit. I threw this together over
the last hour or two. It isn't actually writing to the new files
atm and I am not seeing why so I am hoping a new set of eyes can
see and fix it fast!